Add a basic progress form
