+++ opentech/apply/funds/templates/funds/applicationsubmission_detail.html
@@ -50,6 +50,7 @@
         </div>
 
         <aside class="sidebar">
+            {% include "funds/includes/progress_form.html" %}
             {% if other_submissions %}
                 <div class="sidebar__inner">
                 <h6 class="heading heading--light-grey heading--small heading--uppercase">Past Submissions</h6>
