+++ opentech/apply/funds/templates/funds/includes/progress_form.html
@@ -1,0 +1,5 @@
+<form method="post">
+    {% csrf_token %}
+    {{ progress_form }}
+    <input type="submit" value="Progress">
+</form>
